Adding local DAO for rests

+++ ice/src/main/java/by/ingman/ice/retailerrequest/v2/remote/dao/ProductDao.java
@@ -43,7 +43,7 @@ public class ProductDao {
                             rs.getString("name_p"),
                             rs.getString("packs"),
                             rs.getString("amount"),
-                            rs.getInt("price"),
+                            rs.getDouble("price"),
                             rs.getInt("amt_in_pack"),
                             rs.getDouble("gross_weight")
                     );
